# docs/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/post/{id}:
    get:
      tags:
      - post-controller
      operationId: getPostById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
...
  /api/post-version/{id}:
    get:
      tags:
      - post-version-controller
      operationId: getPostVersionById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostVersion"
  /api/post-version/post/{postId}:
    get:
      tags:
      - post-version-controller
      operationId: getPostVersionsByPostId
      parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostVersion"
components:
  schemas:
...
    User:
      required:
      - createdAt
      - email
      - id
      - name
      - updatedAt
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Post:
      required:
      - content
      - createdAt
      - id
      - updatedAt
      - user
      type: object
      properties:
        id:
          type: integer
          format: int64
        content:
          type: string
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PostVersion:
      required:
      - content
      - createdAt
      - id
      - post
      - updatedAt
      type: object
      properties:
        id:
          type: integer
          format: int64
        post:
          $ref: "#/components/schemas/Post"
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time